examples/o1_web_crawler_actions/requirements.txt
firecrawl-py
python-dotenv
openai